refactor(stacks): replace sidebar nav with top-nav dropdown + overview page create button
Stack navigation is now handled by the hover dropdown on the top nav bar,
making the StackNav sidebar redundant. Removed it and moved stack creation
into the /stacks overview page, which now has a "New stack" button in both
the populated and empty states.
